requirements.txt
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
DateTime==5.2
telegram==0.0.1 # так ли это ?
python-telegram-bot==13.15

